+++ Runtime/Schemas/ArenaConeJson.cs
@@ -32,7 +32,7 @@ namespace ArenaUnity.Schemas
         public float Height = defHeight;
         public bool ShouldSerializeHeight()
         {
-            return true; // required in json schema 
+            return true; // required in json schema
         }
 
         private static bool defOpenEnded = false;
@@ -51,7 +51,7 @@ namespace ArenaUnity.Schemas
         public float RadiusBottom = defRadiusBottom;
         public bool ShouldSerializeRadiusBottom()
         {
-            return true; // required in json schema 
+            return true; // required in json schema
         }
 
         private static float defRadiusTop = 0.01f;
@@ -64,20 +64,20 @@ namespace ArenaUnity.Schemas
             return (RadiusTop != defRadiusTop);
         }
 
-        private static float defSegmentsHeight = 18f;
+        private static int defSegmentsHeight = 18;
         [JsonProperty(PropertyName = "segmentsHeight")]
         [Tooltip("segments height")]
-        public float SegmentsHeight = defSegmentsHeight;
+        public int SegmentsHeight = defSegmentsHeight;
         public bool ShouldSerializeSegmentsHeight()
         {
             if (_token != null && _token.SelectToken("segmentsHeight") != null) return true;
             return (SegmentsHeight != defSegmentsHeight);
         }
 
-        private static float defSegmentsRadial = 36f;
+        private static int defSegmentsRadial = 36;
         [JsonProperty(PropertyName = "segmentsRadial")]
         [Tooltip("segments radial")]
-        public float SegmentsRadial = defSegmentsRadial;
+        public int SegmentsRadial = defSegmentsRadial;
         public bool ShouldSerializeSegmentsRadial()
         {
             if (_token != null && _token.SelectToken("segmentsRadial") != null) return true;
