+++ Runtime/Schemas/ArenaImageJson.cs
@@ -32,7 +32,7 @@ namespace ArenaUnity.Schemas
         public string Url = defUrl;
         public bool ShouldSerializeUrl()
         {
-            return true; // required in json schema 
+            return true; // required in json schema
         }
 
         private static float defHeight = 1f;
@@ -45,20 +45,20 @@ namespace ArenaUnity.Schemas
             return (Height != defHeight);
         }
 
-        private static float defSegmentsHeight = 1f;
+        private static int defSegmentsHeight = 1;
         [JsonProperty(PropertyName = "segmentsHeight")]
         [Tooltip("segments height")]
-        public float SegmentsHeight = defSegmentsHeight;
+        public int SegmentsHeight = defSegmentsHeight;
         public bool ShouldSerializeSegmentsHeight()
         {
             if (_token != null && _token.SelectToken("segmentsHeight") != null) return true;
             return (SegmentsHeight != defSegmentsHeight);
         }
 
-        private static float defSegmentsWidth = 1f;
+        private static int defSegmentsWidth = 1;
         [JsonProperty(PropertyName = "segmentsWidth")]
         [Tooltip("segments width")]
-        public float SegmentsWidth = defSegmentsWidth;
+        public int SegmentsWidth = defSegmentsWidth;
         public bool ShouldSerializeSegmentsWidth()
         {
             if (_token != null && _token.SelectToken("segmentsWidth") != null) return true;
